[refac] rename class
