+++ src/main/java/org/hankki/hankkiserver/api/auth/controller/AuthController.java
@@ -4,7 +4,7 @@ import jakarta.annotation.Nullable;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.hankki.hankkiserver.api.auth.controller.request.UserLoginRequest;
-import org.hankki.hankkiserver.api.auth.service.AuthService;
+import org.hankki.hankkiserver.api.auth.service.AuthFacade;
 import org.hankki.hankkiserver.api.auth.service.response.UserLoginResponse;
 import org.hankki.hankkiserver.api.auth.service.response.UserReissueResponse;
 import org.hankki.hankkiserver.api.common.annotation.UserId;
@@ -24,31 +24,31 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1")
 public class AuthController {
 
-    private final AuthService authService;
+    private final AuthFacade authFacade;
 
     @PostMapping("/auth/login")
     public HankkiResponse<UserLoginResponse> login(@RequestHeader(HttpHeaders.AUTHORIZATION) final String token,
                                                    @Valid @RequestBody final UserLoginRequest request) {
-        UserLoginResponse response = authService.login(token, request);
+        UserLoginResponse response = authFacade.login(token, request);
         return HankkiResponse.success(CommonSuccessCode.OK, response);
     }
 
     @PatchMapping("/auth/logout")
     public HankkiResponse<Void> signOut(@UserId final Long userId) {
-        authService.logout(userId);
+        authFacade.logout(userId);
         return HankkiResponse.success(CommonSuccessCode.OK);
     }
 
     @DeleteMapping("/auth/withdraw")
     public HankkiResponse<Void> withdraw(@UserId final Long userId,
                                          @Nullable @RequestHeader("X-Apple-Code") final String code) {
-        authService.withdraw(userId, code);
+        authFacade.withdraw(userId, code);
         return HankkiResponse.success(CommonSuccessCode.NO_CONTENT);
     }
 
     @PostMapping("/auth/reissue")
     public HankkiResponse<UserReissueResponse> reissue(@RequestHeader(HttpHeaders.AUTHORIZATION) final String token) {
-        UserReissueResponse response = authService.reissue(token);
+        UserReissueResponse response = authFacade.reissue(token);
         return HankkiResponse.success(CommonSuccessCode.OK, response);
     }
 }
